requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
pyodbc==5.0.1
python-dotenv==1.0.1
email-validator==2.1.0.post1
Pillow==10.2.0
Flask-Mail==0.9.1 